Refactored the gang component and created add fighter and add gang addition components
